bot/requirements.txt
aiogram==3.13.1
aiohttp==3.9.5
python-dotenv==1.0.1
httpx==0.27.0
